// src/app/models/user.model.ts
export default class User {
  _userId: bigint;
  devName: string;
  officeCode: string;
  positionName: string;
  nickname: string;
  firstname: string;
  middlename: string;
  lastname: string;
  suffix: string;
  username: string;
  password: string;
  email: string;
  gender: number;
  authority: number;
}